+++ src/apis/service/index.ts
@@ -1,2 +1,3 @@
 export { default as flower } from './flower.ts'
 export { default as letterSent } from './letterSent.ts'
+export { default as writeLetter } from './writeLetter.ts'
